requirements.txt
hf_transfer==0.1.9
openai==1.66.5
nvidia-nvtx-cu12==12.4.127
aiosignal==1.3.2
socksio==1.0.0
outlines==0.1.11
tenacity==9.0.0
fonttools==4.56.0
pydub==0.25.1
httptools==0.6.4
language_data==1.3.0
huggingface-hub==0.29.3
setuptools==75.8.2
pyparsing==3.2.1
urllib3==2.3.0
ninja==1.11.1.3
opencv-python-headless==4.11.0.86
cachetools==5.5.2
colorama==0.4.6
dill==0.3.9
scikit-learn==1.6.1
nvidia-cusparse-cu12==12.3.1.170
sentencepiece==0.2.0
protobuf==6.30.1
websockets==11.0.3
filelock==3.18.0
wheel==0.45.1
srsly==2.5.1
rich-toolkit==0.13.2
blake3==1.0.4
triton==3.2.0
nvidia-cublas-cu12==12.4.5.8
gradio==3.50.2
vllm==0.8.0
RapidFuzz==3.12.2
mpmath==1.3.0
Jinja2==3.1.6
lm-format-enforcer==0.10.11
pandas==2.2.3
cycler==0.12.1
py-cpuinfo==9.0.0
yarl==1.18.3
preshed==3.0.9
python-dateutil==2.9.0.post0
fsspec==2025.3.0
torch==2.6.0
blis==1.2.0
airportsdata==20250224
diskcache==5.6.3
ray==2.43.0
python-multipart==0.0.20
nvidia-cuda-cupti-cu12==12.4.127
aiohappyeyeballs==2.6.1
jsonschema-specifications==2024.10.1
PyYAML==6.0.2
toolz==1.0.0
more-itertools==10.6.0
contourpy==1.3.1
rpds-py==0.23.1
py3langid==0.2.2
msgspec==0.19.0
nvidia-curand-cu12==10.3.5.147
orjson==3.10.15
cupy-cuda12x==13.4.0
pip==25.0.1
python-dotenv==1.0.1
aiofiles==23.2.1
torchvision==0.21.0
inflect==7.5.0
einops==0.8.1
aiohttp==3.11.14
psutil==7.0.0
threadpoolctl==3.6.0
typing_extensions==4.12.2
Pygments==2.19.1
multidict==6.2.0
mdurl==0.1.2
colorlog==6.9.0
astor==0.8.1
transformers==4.49.0
prometheus_client==0.21.1
nvidia-cudnn-cu12==9.1.0.70
typer==0.15.2
pytz==2025.1
wrapt==1.17.2
watchfiles==1.0.4
wasabi==1.1.3
aiosignal==1.3.2
attrs==25.1.0
aiohttp==3.11.13
multidict==6.1.0
yarl==1.18.3
propcache==0.3.0
idna==3.10
aiohappyeyeballs==2.4.6
frozenlist==1.5.0
setproctitle==1.2.2
psutil==7.0.0
colorama==0.4.6
nest-asyncio==1.6.0
distro==1.9.0
Antergos==Version:
torchaudio==2.6.0
zipp==3.21.0
idna==3.10
langcodes==3.5.0
annotated-types==0.7.0
smart-open==7.1.0
frozenlist==1.5.0
nvidia-cuda-runtime-cu12==12.4.127
murmurhash==1.0.12
prometheus-fastapi-instrumentator==7.0.2
pydantic_core==2.27.2
outlines_core==0.1.26
referencing==0.36.2
httpcore==0.18.0
compressed-tensors==0.9.2
depyf==0.18.0
gguf==0.10.0
python-json-logger==3.3.0
mistral_common==1.5.4
sympy==1.13.1
numba==0.60.0
spacy-loggers==1.0.5
nvidia-nvjitlink-cu12==12.4.127
tokenizers==0.21.1
partial-json-parser==0.2.1.1.post5
safetensors==0.5.3
cloudpathlib==0.21.0
gradio_client==0.6.1
nvidia-nccl-cu12==2.21.5
joblib==1.4.2
nvidia-cuda-nvrtc-cu12==12.4.127
catalogue==2.0.10
pillow==10.4.0
kiwisolver==1.4.8
propcache==0.3.0
tqdm==4.67.1
importlib_metadata==8.6.1
altair==5.3.0
xgrammar==0.1.16
thinc==8.3.4
backports.tarfile==1.2.0
more-itertools==10.3.0
tomli==2.0.1
jaraco.collections==5.1.0
packaging==24.2
importlib_metadata==8.0.0
platformdirs==4.2.2
wheel==0.43.0
typing_extensions==4.12.2
zipp==3.19.2
jaraco.context==5.3.0
jaraco.text==3.12.1
autocommand==2.2.2
inflect==7.3.1
jaraco.functools==4.0.1
typeguard==4.3.0
click==8.1.8
charset-normalizer==3.4.1
opencv-python==4.8.0.76
pyzmq==26.3.0
jiter==0.9.0
semantic-version==2.10.0
en_core_web_sm==3.8.0
ffmpy==0.5.0
spacy-legacy==3.0.12
six==1.17.0
nvidia-cusparselt-cu12==0.6.2
uvloop==0.21.0
packaging==24.2
httpx==0.25.0
cloudpickle==3.1.1
shellingham==1.5.4
nltk==3.9.1
confection==0.1.5
typing-inspect==0.8.0
certifi==2025.1.31
pydantic==2.10.6
sniffio
==1.3.1
numpy==1.26.4
pycountry==24.6.1
cymem==2.0.11
MarkupSafe==2.1.5
requests==2.32.3
email_validator==2.2.0
nvidia-cusolver-cu12==11.6.1.9
setuptools-scm==8.2.0
dnspython==2.7.0
marisa-trie==1.2.1
typeguard==4.4.2
networkx==3.4.2
h11==0.14.0
matplotlib==3.10.1
spacy==3.8.4
weasel==0.4.1
fastrlock==0.8.3
attrs==25.3.0
tiktoken==0.9.0
markdown-it-py==3.0.0
mypy-extensions==1.0.0
nvidia-cufft-cu12==11.2.1.3
uvicorn==0.34.0
scipy==1.15.2
rich==13.9.4
fastapi==0.115.11
lark==1.2.2
tzdata==2025.1
fastapi-cli==0.0.7
anyio==3.7.1
interegular
==0.3.3
my-test-package==1.0
xformers==0.0.29.post2
jsonschema==4.23.0
starlette==0.46.1
msgpack==1.1.0
astunparse==1.6.3
llvmlite==0.43.0
importlib_resources==6.5.2
regex==2024.11.6
